requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=0.8.11
plotly>=5.15.0
docx2pdf>=0.1.8
Pillow>=10.0.0
xlsxwriter>=3.1.0
